+++ SportAssociation/SportAssociation/Controllers/Host_RequestController.cs
@@ -22,45 +22,64 @@ namespace SportAssociation.Controllers
         // GET: Host_Request
         public async Task<IActionResult> Index()
         {
-            var outputSQLParam = new Microsoft.Data.SqlClient.SqlParameter("@out", System.Data.SqlDbType.Int) { Direction = System.Data.ParameterDirection.Output };
-            _context.Database.ExecuteSqlRaw("exec dbo.getManagerID3 @username='" + Authentication.username + "', @out={0} out",
-                outputSQLParam);
-
-            int output  = -1;
-
-            if (outputSQLParam.Value != DBNull.Value)
+            if (Super_UserController.currentUser == "Manager")
             {
-                output = (int)outputSQLParam.Value;
+                var outputSQLParam = new Microsoft.Data.SqlClient.SqlParameter("@out", System.Data.SqlDbType.Int) { Direction = System.Data.ParameterDirection.Output };
+                _context.Database.ExecuteSqlRaw("exec dbo.getManagerID3 @username='" + Authentication.username + "', @out={0} out",
+                    outputSQLParam);
+
+                int output = -1;
+
+                if (outputSQLParam.Value != DBNull.Value)
+                {
+                    output = (int)outputSQLParam.Value;
+                }
+
+                ViewBag.currentManagerID = output;
+
+
+                return View(await _context.Host_Request.ToListAsync());
             }
-
-            ViewBag.currentManagerID = output;
-
-
-            return View(await _context.Host_Request.ToListAsync());
+            else
+            {
+                TempData["alertMessage"] = "You can not access this page.";
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: Host_Request/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null || _context.Host_Request == null)
+            if (Super_UserController.currentUser == "Manager")
             {
-                return NotFound();
-            }
+                if (id == null || _context.Host_Request == null)
+                {
+                    return NotFound();
+                }
 
-            var host_Request = await _context.Host_Request
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (host_Request == null)
+                var host_Request = await _context.Host_Request
+                    .FirstOrDefaultAsync(m => m.Id == id);
+                if (host_Request == null)
+                {
+                    return NotFound();
+                }
+
+                return View(host_Request);
+            }
+            else
             {
-                return NotFound();
+                TempData["alertMessage"] = "You can not access this page.";
+                return RedirectToAction("Index", "Home");
             }
-
-            return View(host_Request);
         }
 
         // GET: Host_Request/Create
         public IActionResult Create()
         {
-            return View();
+            if(Super_UserController.currentUser == "")
+                return View();
+            TempData["alertMessage"] = "You can not access this page.";
+            return RedirectToAction("Index", "Home");
         }
 
         // POST: Host_Request/Create
@@ -70,29 +89,45 @@ namespace SportAssociation.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,representative_id,manager_id,match_id,status")] Host_Request host_Request)
         {
-            if (ModelState.IsValid)
+            if (Super_UserController.currentUser == "")
             {
-                _context.Add(host_Request);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                if (ModelState.IsValid)
+                {
+                    _context.Add(host_Request);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
+                }
+                return View(host_Request);
             }
-            return View(host_Request);
+            else
+            {
+                TempData["alertMessage"] = "You can not access this page.";
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: Host_Request/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null || _context.Host_Request == null)
+            if (Super_UserController.currentUser == "Manager")
             {
-                return NotFound();
-            }
+                if (id == null || _context.Host_Request == null)
+                {
+                    return NotFound();
+                }
 
-            var host_Request = await _context.Host_Request.FindAsync(id);
-            if (host_Request == null)
-            {
-                return NotFound();
+                var host_Request = await _context.Host_Request.FindAsync(id);
+                if (host_Request == null)
+                {
+                    return NotFound();
+                }
+                return View(host_Request);
             }
-            return View(host_Request);
+            else
+            {
+                TempData["alertMessage"] = "You can not access this page.";
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // POST: Host_Request/Edit/5
@@ -102,67 +137,83 @@ namespace SportAssociation.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,representative_id,manager_id,match_id,status")] Host_Request host_Request)
         {
-            if (id != host_Request.Id)
+            if (Super_UserController.currentUser == "Manager")
             {
-                return NotFound();
-            }
+                if (id != host_Request.Id)
+                {
+                    return NotFound();
+                }
 
-            if (ModelState.IsValid)
-            {
-                try
+                if (ModelState.IsValid)
                 {
-                    _context.Update(host_Request);
-                    await _context.SaveChangesAsync();
-                    if (host_Request.status == true)
+                    try
                     {
-                        _context.Database.ExecuteSqlRaw("delete from Host_Request where match_id = '" + host_Request.match_id + "' and status is null");
-                        _context.Database.ExecuteSqlRaw("update Match set stadium_id = (SELECT stadium_id FROM Manager where id = '" + host_Request.manager_id + "') where id = '" + host_Request.match_id + "'");
-                        var outputSQLParam = new Microsoft.Data.SqlClient.SqlParameter("@out", System.Data.SqlDbType.Int) { Direction = System.Data.ParameterDirection.Output };
-                        _context.Database.ExecuteSqlRaw("exec dbo.getCapacity @manager_id='" + host_Request.manager_id + "', @out={0} out",
-                            outputSQLParam);
-                        int output = -1;
-                        if (outputSQLParam.Value != DBNull.Value)
+                        _context.Update(host_Request);
+                        await _context.SaveChangesAsync();
+                        if (host_Request.status == true)
                         {
-                            output = (int)outputSQLParam.Value;
+                            _context.Database.ExecuteSqlRaw("delete from Host_Request where match_id = '" + host_Request.match_id + "' and status is null");
+                            _context.Database.ExecuteSqlRaw("update Match set stadium_id = (SELECT stadium_id FROM Manager where id = '" + host_Request.manager_id + "') where id = '" + host_Request.match_id + "'");
+                            var outputSQLParam = new Microsoft.Data.SqlClient.SqlParameter("@out", System.Data.SqlDbType.Int) { Direction = System.Data.ParameterDirection.Output };
+                            _context.Database.ExecuteSqlRaw("exec dbo.getCapacity @manager_id='" + host_Request.manager_id + "', @out={0} out",
+                                outputSQLParam);
+                            int output = -1;
+                            if (outputSQLParam.Value != DBNull.Value)
+                            {
+                                output = (int)outputSQLParam.Value;
+                            }
+                            for (int i = 0; i < output; i++)
+                            {
+                                _context.Database.ExecuteSqlRaw("Insert into Ticket values(1, " + host_Request.match_id + ")");
+                            }
                         }
-                        for(int i=0; i < output; i++)
+                    }
+                    catch (DbUpdateConcurrencyException)
+                    {
+                        if (!Host_RequestExists(host_Request.Id))
                         {
-                            _context.Database.ExecuteSqlRaw("Insert into Ticket values(1, " + host_Request.match_id + ")");
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
                         }
                     }
+                    return RedirectToAction(nameof(Index));
                 }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!Host_RequestExists(host_Request.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
+                return View(host_Request);
             }
-            return View(host_Request);
+            else
+            {
+                TempData["alertMessage"] = "You can not access this page.";
+                return RedirectToAction("Index", "Home");
+            }
         }
 
         // GET: Host_Request/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null || _context.Host_Request == null)
+            if (Super_UserController.currentUser == "Manager")
             {
-                return NotFound();
-            }
+                if (id == null || _context.Host_Request == null)
+                {
+                    return NotFound();
+                }
 
-            var host_Request = await _context.Host_Request
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (host_Request == null)
+                var host_Request = await _context.Host_Request
+                    .FirstOrDefaultAsync(m => m.Id == id);
+                if (host_Request == null)
+                {
+                    return NotFound();
+                }
+
+                return View(host_Request);
+            }
+            else
             {
-                return NotFound();
+                TempData["alertMessage"] = "You can not access this page.";
+                return RedirectToAction("Index", "Home");
             }
-
-            return View(host_Request);
         }
 
         // POST: Host_Request/Delete/5
@@ -170,18 +221,26 @@ namespace SportAssociation.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Host_Request == null)
+            if (Super_UserController.currentUser == "Manager")
             {
-                return Problem("Entity set 'ApplicationDbContext.Host_Request'  is null.");
+                if (_context.Host_Request == null)
+                {
+                    return Problem("Entity set 'ApplicationDbContext.Host_Request'  is null.");
+                }
+                var host_Request = await _context.Host_Request.FindAsync(id);
+                if (host_Request != null)
+                {
+                    _context.Host_Request.Remove(host_Request);
+                }
+
+                await _context.SaveChangesAsync();
+                return RedirectToAction(nameof(Index));
             }
-            var host_Request = await _context.Host_Request.FindAsync(id);
-            if (host_Request != null)
+            else
             {
-                _context.Host_Request.Remove(host_Request);
+                TempData["alertMessage"] = "You can not access this page.";
+                return RedirectToAction("Index", "Home");
             }
-            
-            await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
         }
 
         private bool Host_RequestExists(int id)
